Created the "Content.js" component
The component displays the description and image of given news. It is imported into "Index.js" to be rendered there.
