+++ src/App.js
@@ -13,7 +13,7 @@ import SchoolCampaign from './Campaigns/School/SchoolCampaign';
 import ToysCampaign from './Campaigns/Toys/ToysCampaign';
 import MembersForm from './Components/Members/MembersForm';
 import ProjectsForm from './Components/Projects/ProjectsForm';
-import NewsDetail from './Components/News/Detail';
+import NewsDetail from './Components/News/Detail/Index';
 
 function App() {
   return (
